Address feedback & do misc clean up
