+++ src/libraries/Common/src/Roslyn/DiagnosticDescriptorHelper.cs
@@ -1,6 +1,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.CodeAnalysis;
+
 namespace Microsoft.CodeAnalysis.DotnetRuntime.Extensions
 {
     internal static partial class DiagnosticDescriptorHelper
