+++ src/libraries/Microsoft.Extensions.Configuration.Binder/gen/Microsoft.Extensions.Configuration.Binder.SourceGeneration.csproj
@@ -47,7 +47,6 @@
     <Compile Include="Parser\ConfigurationBinder.cs" />
     <Compile Include="Parser\Extensions.cs" />
     <Compile Include="Parser\DiagnosticDescriptors.cs" />
-    <Compile Include="Parser\Helpers.cs" />
     <Compile Include="Parser\KnownTypeSymbols.cs" />
     <Compile Include="Parser\OptionsBuilderConfigurationExtensions.cs" />
     <Compile Include="Parser\OptionsConfigurationServiceCollectionExtensions.cs" />
