+++ src/libraries/System.Text.Json/gen/Helpers/RoslynExtensions.cs
@@ -40,6 +40,61 @@ namespace System.Text.Json.SourceGeneration
         public static bool ContainsLocation(this Compilation compilation, Location location)
             => location.SourceTree != null && compilation.ContainsSyntaxTree(location.SourceTree);
 
+        /// <summary>
+        /// Removes any type metadata that is erased at compile time, such as NRT annotations and tuple labels.
+        /// </summary>
+        public static ITypeSymbol EraseCompileTimeMetadata(this Compilation compilation, ITypeSymbol type)
+        {
+            if (type.NullableAnnotation is NullableAnnotation.Annotated)
+            {
+                type = type.WithNullableAnnotation(NullableAnnotation.None);
+            }
+
+            if (type is INamedTypeSymbol namedType)
+            {
+                if (namedType.IsTupleType)
+                {
+                    if (namedType.TupleElements.Length < 2)
+                    {
+                        return type;
+                    }
+
+                    ImmutableArray<ITypeSymbol> erasedElements = namedType.TupleElements
+                        .Select(e => compilation.EraseCompileTimeMetadata(e.Type))
+                        .ToImmutableArray();
+
+                    type = compilation.CreateTupleTypeSymbol(erasedElements);
+                }
+                else if (namedType.IsGenericType)
+                {
+                    if (namedType.IsUnboundGenericType)
+                    {
+                        return namedType;
+                    }
+
+                    ImmutableArray<ITypeSymbol> typeArguments = namedType.TypeArguments;
+                    INamedTypeSymbol? containingType = namedType.ContainingType;
+
+                    if (containingType?.IsGenericType == true)
+                    {
+                        containingType = (INamedTypeSymbol)compilation.EraseCompileTimeMetadata(containingType);
+                        type = namedType = containingType.GetTypeMembers().First(t => t.Name == namedType.Name && t.Arity == namedType.Arity);
+                    }
+
+                    if (typeArguments.Length > 0)
+                    {
+                        ITypeSymbol[] erasedTypeArgs = typeArguments
+                            .Select(compilation.EraseCompileTimeMetadata)
+                            .ToArray();
+
+                        type = namedType.ConstructedFrom.Construct(erasedTypeArgs);
+                    }
+                }
+            }
+
+            return type;
+        }
+
         public static bool CanUseDefaultConstructorForDeserialization(this ITypeSymbol type, out IMethodSymbol? constructorInfo)
         {
             if (type.IsAbstract || type.TypeKind is TypeKind.Interface || type is not INamedTypeSymbol namedType)
